+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+
+- The mandatory docstring gate could skip its own scan and still exit zero ([pm-jira-5vse](https://github.com/unbraind/pm-jira/blob/main/.agents/pm/issues/pm-jira-5vse.toon))
+
+### Other
+
+- Adopt the canonical pm-ops docstring gate ([pm-jira-j6sr](https://github.com/unbraind/pm-jira/blob/main/.agents/pm/tasks/pm-jira-j6sr.toon))
+
 ## 2026.8.7 - 2026-08-07
 
 ### Other